backend/requirements.txt
# FastAPI & ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
motor>=3.3.2  # Async MongoDB driver
pymongo>=4.9

# Task Queue
celery>=5.3.4
redis>=5.0.1

# AI SDKs
openai>=1.3.7
anthropic>=0.7.7

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0
python-multipart>=0.0.6

# Document generation
python-docx>=1.1.0

# NLP utilities for word counting
# spacy==3.7.2  # Commented out - requires C++ build tools on Windows, not needed for outline generation

# YAML parsing
PyYAML>=6.0.1

# HTTP client
httpx>=0.25.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
fakeredis>=2.20.0

# Code quality
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.1
